fix(plugins): guard reconciliation against fabricated disputes on agreement
Both the state and scalar-field reconcilers were missing an equality check:
when local and remote independently arrived at the same value, the missing
guard let them report it as a dispute anyway. Add the guard to both
(findings #1/#8 state.rs, #2/#8 fields.rs), and switch the label
write-decision from an order-sensitive vector comparison to a set
comparison (finding #9) so re-ordered-but-identical label arrays stop
triggering perpetual PATCHes.

reconcile_tests.rs crossed the 300-line cap once the new dispute/order
cases were added; split it into reconcile_tests/{fields,labels,state}.rs
along the same seams as reconcile/ itself (finding #13).
